+++ 03.개발/cbe/src/kr/co/studystory/mapper/study_group_mapper3.xml
@@ -24,4 +24,18 @@
 		</where>
 		ORDER BY input_date DESC, accept_flag desc, deactivation
 	</select>
+	
+	  
+  <select id="selectMyStudies" resultMap="myStudyResult" parameterType="String">
+	select s.s_num, study_name, loc, category, s.img, recruitment, input_date, s.deactivation, u.nick
+	from study s, user_table u, member m
+	where u.id=#{ id } and m.id=u.id and m.s_num=s.s_num;
+  </select>
+  
+  <select id="selectAppliedStudies" resultMap="appliedStudyResult" parameterType="String">
+	select s.s_num, study_name, loc, category, s.img, recruitment, s.input_date, s.deactivation, u.nick
+	from user_table u, study s, join j
+	where u.id=#{ id } and j.id = u.id and j.s_num = s.s_num;
+  </select>
+  
 </mapper>
